final_version1/requirements.txt
google-api-python-client==1.6.6
flask_table
Flask-WTF

